requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.4
python-dotenv==1.0.1
twilio==8.12.0
django-rest-knox==4.2.0
Pillow==10.2.0
requests==2.31.0 